Few things with delta Time

+++ src/framework/rendering/Time.java
@@ -1,5 +1,5 @@
 package framework.rendering;
 
 public class Time {
-static public float deltaTime ; //time since last frame in seconds
+static public double deltaTime ; //time since last frame in seconds
 }
